{{cookiecutter.directory_name}}/requirements.txt
Django==3.2.13
django-cors-headers==3.13.0
django-debug-toolbar==3.2.2
django-environ==0.9.0
django-extensions==3.1.3
django-fsm==2.7.1
django-fsm-admin==1.2.4
django-guardian==2.4.0
acdh-django-browsing==1.0.0
django-spaghetti-and-meatballs==0.4.2
django-tabbed-admin==1.0.4
requests>=2.25
{% if cookiecutter.ldap_authentication == 'y' -%}
python-ldap==3.3.1
django-auth-ldap==3.0.0
{%- endif %}